requirements.txt
tqdm==4.65.0
python-dotenv==1.0.1
openai==1.13.3
requests==2.31.0
numpy==1.26.4
pandas==1.5.2